requirements.txt
Django==3.2.7
dj-database-url==0.5.0
django-on-heroku==1.1.2
django-environ==0.8.1
espn-api==0.17.0
gunicorn==20.1.0
numpy==1.17.5
pandas==0.25.3
psycopg2-binary==2.9.1
tabulate==0.8.9
tweepy==4.1.0
nltk==3.6.1
wordcloud==1.8.1
scikit-learn==0.22.2.post1
scikit-image==0.18.1